% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcplot.R
\name{gcplot}
\alias{gcplot}
\title{Convert a ggplot into a Good Charts themed plot}
\usage{
gcplot(plot, title = NA, subtitle = NA, ylab = NA, caption = NA)
}
\arguments{
\item{plot}{A ggplot object}

\item{title}{(optional) The plot's title}

\item{subtitle}{(optional) The plot's subtitle}

\item{ylab}{(optional) The y-axis label}

\item{caption}{(optional) The caption or source line}
}
\description{
Convert a ggplot into a Good Charts themed plot
}
\examples{
\dontrun{
library(ggplot2)

p1 <- ggplot(iris, aes(Petal.Length, Petal.Width, color = Species)) +
  geom_point() +
  labs(
    title = "A ROSE BY ANY OTHER NAME",
    subtitle = "IRISES PETAL ATTRIBUTE",
    caption = "SOURCE: R.A. FISHER",
    x = "PETAL LENGTH",
    y = "PETAL WIDTH"
  )
gcplot(p1)
}
}
