% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gc_font.R
\name{get_gc_font}
\alias{get_gc_font}
\title{Get a Good Charts style font}
\usage{
get_gc_font(chart_element = NA)
}
\arguments{
\item{chart_element}{The best font for the chart element}
}
\description{
Get a Good Charts style font
}
