% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gc_fonts.R
\name{setup_gc_fonts}
\alias{setup_gc_fonts}
\title{Setup fonts for Good Charts Theme}
\usage{
setup_gc_fonts(import_fonts = FALSE)
}
\arguments{
\item{import_fonts}{boolean if the fonts should be imported}
}
\description{
Setup fonts for Good Charts Theme
}
