% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gc.R
\name{theme_gc}
\alias{theme_gc}
\alias{theme_good_charts}
\title{Good Charts Theme}
\usage{
theme_gc(...)
}
\arguments{
\item{...}{All arguments to pass into the base ggplot theme}
}
\description{
Good Charts Theme
}
\examples{
\dontrun{
library(ggplot2)

ggplot(iris, aes(Petal.Length, Petal.Width, color = Species)) +
  geom_point() +
  theme_gc()
}
}
