% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_chart.R
\name{unit_chart}
\alias{unit_chart}
\title{Good Charts Inspired Unit Chart}
\usage{
unit_chart(
  n,
  out_of,
  color_random_points = FALSE,
  highlight_color = "#E41A1C",
  other_color = "#DFDFDF"
)
}
\arguments{
\item{n}{number of items to be colored}

\item{out_of}{Total number of units}

\item{color_random_points}{(optional) Highlight random points}

\item{highlight_color}{(optional) The highlight point color}

\item{other_color}{(optional) The color for all the other points}
}
\description{
Good Charts Inspired Unit Chart
}
\examples{
# Create the basketball chart on page 145
unit_chart(3, 10000, TRUE)
}
